Move function resolvers in a separate file
